pyproject.toml
[project]
name = "rembrandt-chat"
version = "0.2.0"
description = "A Client for chat sessions of Rembrandt."
authors = [{ name = "Carlos Pinto", email = "carlospinto4p@gmail.com" }]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "rembrandt @ git+https://github.com/carlospinto4p/rembrandt.git@v2.6.0",
    "python-telegram-bot>=21.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.12.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
]

[tool.ruff]
line-length = 78
